Updated to All wallets and incooperated PR changes
