+++ core2/pkg/accounting/00_module.go
@@ -9,6 +9,5 @@ func Init() {
 	initSupportAssistAcc()
 	initUsageReports()
 	initUsageGenerator()
-
-	go InitLowFundsScan()
+	initLowFundsScan()
 }
